Fix clang compilation error (#1242)
min(int, size_t) is ambiguous

[ROCm/hipBLASLt commit: 1342b7a870ff16107a0248602e4655addeb04965]

+++ library/src/amd_detail/rocblaslt/src/tensile_host.cpp
@@ -2895,7 +2895,7 @@ rocblaslt_status getBestSolutions(rocblaslt_handle       handle,
                                      requestedAlgoCount);
         }
 
-        auto algoCount       = min(requestedAlgoCount, solutions.size());
+        auto algoCount       = min(static_cast<size_t>(requestedAlgoCount), solutions.size());
         int  returnAlgoCount = 0;
         heuristicResults.clear();
         heuristicResults.resize(algoCount);
@@ -2920,7 +2920,7 @@ rocblaslt_status getBestSolutions(rocblaslt_handle       handle,
         auto solutions = library->findTopSolutionsGroupedGemm(
             data->problem.gemms, *hardware, requestedAlgoCount);
 
-        auto algoCount       = min(requestedAlgoCount, solutions.size());
+        auto algoCount       = min(static_cast<size_t>(requestedAlgoCount), solutions.size());
         int  returnAlgoCount = 0;
         heuristicResults.clear();
         heuristicResults.resize(algoCount);
